+++ lib/eq_parser.dart
@@ -1,6 +1,6 @@
 /// Parse math equations from a string
 ///
-/// As well as mathematical operations, has support for variables and functions.
+/// Has support for mathematical operations, variables and functions.
 library eq_parser;
 
 export 'src/eq_parser.dart';